packages/forecasting/requirements.txt
torch
pytorch-forecasting
pytorch-lightning
pandas
numpy
pytest
scikit-learn
